+++ tests/core/test_storage.py
@@ -1150,6 +1150,15 @@ class StorageTest(ThreadMixin,
     """Compound of all storage tests."""
     pass
 
+    def test_records_filtered_when_searched_by_string_field(self):
+        self.create_record({'name': 'foo'})
+        self.create_record({'name': 'bar'})
+        self.create_record({'name': 'FOOBAR'})
+
+        filters = [Filter('name', 'FoO', utils.COMPARISON.LIKE)]
+        results, count = self.storage.get_all(filters=filters, **self.storage_kw)
+        self.assertEqual(len(results), 2)
+
 
 class MemoryStorageTest(StorageTest, unittest.TestCase):
     backend = memory